+++ src/ddsviewprovider.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
-import { ddsToPng } from './util/image/converter';
+import { ddsToPng, tgaToPng } from './util/image/converter';
 import { PNG } from 'pngjs';
 import { localize } from './util/i18n';
 import { DDS } from './util/image/dds';
@@ -17,8 +17,14 @@ export class DDSViewProvider /* implements vscode.CustomEditorProvider */ {
 
     public async resolveCustomEditor(document: { uri: vscode.Uri }, webviewPanel: vscode.WebviewPanel, token: vscode.CancellationToken): Promise<void> {
         try {
-            sendEvent('preview.dds');
             ensureFileScheme(document.uri);
+
+            const documentPath = document.uri.fsPath.toLowerCase();
+            if (documentPath.endsWith('.dds')) {
+                sendEvent('preview.dds');
+            } else {
+                sendEvent('preview.tga');
+            }
 
             const buffer = await Promise.race([
                 fs.promises.readFile(document.uri.fsPath),
@@ -29,9 +35,18 @@ export class DDSViewProvider /* implements vscode.CustomEditorProvider */ {
                 return;
             }
 
-            const dds = DDS.parse(buffer.buffer);
-            const png = ddsToPng(dds);
-            const pngBuffer = PNG.sync.write(png);
+            let png: PNG;
+            let pngBuffer: Buffer;
+            if (documentPath.endsWith('.dds')) {
+                const dds = DDS.parse(buffer.buffer);
+                png = ddsToPng(dds);
+                pngBuffer = PNG.sync.write(png);
+
+            } else {
+                png = tgaToPng(buffer);
+                pngBuffer = PNG.sync.write(png);
+            }
+
             const styleTable = new StyleTable();
 
             webviewPanel.webview.html = html(