+++ src/def.d.ts
@@ -21,3 +21,20 @@ declare module '*.css' {
 
 declare const VERSION: string;
 declare const EXTENSION_ID: string;
+
+declare module 'tga' {
+    class TGA {
+        constructor(buffer: Buffer, opt?: unknown);
+        static createTgaBuffer(width: number, height: number, pixels: [], dontFlipY: boolean): Buffer;
+        static getHeader(buffer: Buffer): unknown;
+        parse(): void;
+        readHeader(): unknown;
+        check(): boolean;
+        addPixel(arr: number[], offset: number, idx: number): void;
+        readPixels(): void;
+        width: number;
+        height: number;
+        pixels: Uint8Array | undefined;
+    }
+    export = TGA;
+}